Extract node builder into its own module
